requirements.txt
alembic==1.8.1
sqlalchemy==1.4.44
pydantic[email]
pyrebase4==4.6.0
firebase-admin==6.0.1
python-dotenv==0.21.0
fastapi==0.88.0
pydantic==1.10.2
uvicorn==0.20.0
h11==0.14.0
click==8.1.3
starlette==0.22.0
asgiref
typing_extensions==4.4.0
anyio==3.6.2
sniffio==1.3.0
psycopg2==2.9.5
google-api-core==2.11.0
google_auth==2.15.0
six==1.16.0
requests==2.28.1
pyasn1==0.4.8
typer==0.7.0
fastapi_migrations==0.0.4
